pyproject.toml
[project]
name = "Companies_Search"
version = "0.1.0"
requires-python = "==3.12.6"
dependencies = [
    "bs4>=0.0.2",
    "pandas>=2.3.1",
    "plotly>=6.3.0",
    "psycopg2-binary==2.9.10",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.48.1",
]
